+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-14T11:49:00.661Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-14T14:13:25.174Z")
     r"""The date and time the resource was last updated"""