+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -101,5 +101,5 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-10-14T11:49:03.658Z")
+    created: Optional[datetime] = parse_datetime("2025-10-14T14:13:28.247Z")
     r"""The date and time the resource was created"""