+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -96,7 +96,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-14T11:49:00.661Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-14T14:13:25.174Z")
     r"""The date and time the resource was last updated"""