+++ packages/orq-rc/RELEASES.md
@@ -3289,3 +3289,13 @@ Based on:
 - [python v3.14.0-rc.24] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.24] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.24 - packages/orq-rc
+
+## 2025-10-14 14:15:08
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.25] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.25] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.25 - packages/orq-rc